Shifted log capture code into preloader script.
