--- interface/i18n/app.de.i18n.json
@@ -1,34 +1,0 @@
-{
-    "app": {
-        "loading": "Lade...",
-        "offline": "Kann mich nicht verbinden, bist du offline?",
-        "logginIn": "Einloggen..."
-    },
-    "error": {
-        "insufficientRights": "Du hast nicht genügend Rechte für diese Aktion."
-    },
-    "buttons": {
-        "ok": "OK",
-        "cancel": "Abbrechen",
-        "save": "Speichern",
-        "edit": "Editieren",
-        "send": "Senden",
-        "next": "Weiter",
-        "previous": "Zurück",
-        "back": "Zurück",
-        "skip": "Überspringen",
-        "sending": "Sende...",
-        "create": "Erstellen",
-        "tryToReconnect": "Versuche mich zu verbinden"
-    },
-    "commonWords": {
-        "you": "Du",
-        "send": "Senden",
-        "or": "oder",
-        "of": "von",
-        "with": "mit",
-        "and": "und",
-        "on": "auf",
-        "per": "pro"
-    }
-}
