+++ modules/ipcCommunicator.js
@@ -7,10 +7,12 @@ Window communication
 const app = require('app');  // Module to control application life.
 const appMenu = require('./menuItems');   
 const popupWindow = require('./popupWindow.js');
-const log = require('./utils/logger').create('ipcCommunicator');
+const logger = require('./utils/logger');
 const ipc = require('electron').ipcMain;
 
 const _ = global._;
+
+const log = logger.create('ipcCommunicator');
 
 /*
 
@@ -136,3 +138,25 @@ ipc.on('backendAction_importPresaleFile', function(e, path, pw) {
 ipc.on('mistAPI_requestAccount', function(e){
     popupWindow.show('requestAccount', {width: 400, height: 230, alwaysOnTop: true}, null, e);
 });
+
+
+
+const uiLoggers = {};
+
+ipc.on('console_log', function(event, id, logLevel, logItemsStr) {
+    try {
+        let loggerId = `(ui: ${id})`;
+
+        let windowLogger = uiLoggers[loggerId];
+
+        if (!windowLogger) {
+            windowLogger = uiLoggers[loggerId] = logger.create(loggerId);
+        }
+
+        windowLogger[logLevel].apply(windowLogger, _.toArray(JSON.parse(logItemsStr)));
+    } catch (err) {
+        log.error(err);
+    }
+});
+
+
